Assign PairID to the last card in shuffledCardsList
PairID is assigned to the last card stored in the list shuffledCardsList, but the value of "newPairID" increments every two times the loop enters to the conditional "!isCardRepeated".

Also I renamed the list "allCards2" to "shuffledCardsList".

Also I removed the "Random Generate PairID" loop.

And I optimized some others things in this script, to look a bit cleaner.